refactor: 테스트에만 사용되는 클래스 테스트 패키지로 이동
