+++ src/main/java/lotto/domain/LottoNumberGenerator.java
@@ -2,6 +2,6 @@ package lotto.domain;
 
 import java.util.Set;
 
-interface LottoNumberGenerator {
+public interface LottoNumberGenerator {
     Set<Integer> generateNumbers();
 }
